backend/orders/requirements.txt
fastapi
uvicorn[standard]
httpx
SQLAlchemy
asyncpg
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-httpx
requests
urllib3
python-dotenv